Add custom uninstall scripts

+++ src/commands/packages/Remove.hx
@@ -1,5 +1,10 @@
 package commands.packages;
 
+import haxe.io.Path;
+import sys.io.File;
+import haxe.Json;
+import structs.Package;
+
 class Remove implements Command {
 	public function bind(args:Array<String>) {
 		var packages = args;
@@ -12,7 +17,17 @@ class Remove implements Command {
 
 			for (pkg in Database.get().packages) {
 				if (pkg.name == pkgname) {
-					Sys.command("rm /usr/bin/" + pkg.name);
+					var zipname = pkg.name + "-" + pkg.version;
+					var packagejson:Package = Json.parse(File.getContent(Path.join(["/opt/escam/temp/", zipname, "package.json"])));
+
+					var uninstallscript = packagejson.scripts.uninstall;
+
+					if (uninstallscript != null) {
+						Sys.command(uninstallscript);
+					} else {
+						Sys.command("rm /usr/bin/" + pkg.name);
+					}
+
 					Sys.println("Removed " + pkgname);
 					Sys.println("Updating database");
 					var db = Database.get();
